requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
google-generativeai